🎨 Remove Superfluous Rule from helm-dashboard ClusterRole
A superfluous rule is added to the ClusterRole upon creation,
when the dashboard.allowWriteActions value is set to true.
This commit will ensure that only a single rule is created
within the ClusterRole, regardless of whether the dashboard.allowWriteActions
value is enabled or not. The verbs within this rule will update accordingly.